+++ src/main/java/com/mercans/integration_api/jpa/repository/EmployeeRepository.java
@@ -1,13 +1,12 @@
 package com.mercans.integration_api.jpa.repository;
 
 import com.mercans.integration_api.jpa.EmployeeEntity;
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Set;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface EmployeeRepository extends JpaRepository<EmployeeEntity, BigInteger> {
+public interface EmployeeRepository extends JpaRepository<EmployeeEntity, Long> {
 
   @Query(" SELECT em.employeeCode FROM EmployeeEntity em ")
   Set<String> getAllEmployeeCodes();
